Split up the workflows into their own files
