+++ workflows/environment_upsert_test.go
@@ -50,44 +50,19 @@ func TestNewEnvironmentUpserter(t *testing.T) {
 	assert.NotNil(upserter)
 }
 
-func TestNewEnvironmentViewer(t *testing.T) {
-	assert := assert.New(t)
-	ctx := common.NewContext()
-	viewer := NewEnvironmentViewer(ctx, "foo", nil)
-	assert.NotNil(viewer)
-}
-
-func TestNewEnvironmentLister(t *testing.T) {
-	assert := assert.New(t)
-	ctx := common.NewContext()
-	lister := NewEnvironmentLister(ctx, nil)
-	assert.NotNil(lister)
-}
-
-func TestNewEnvironmentTerminator(t *testing.T) {
-	assert := assert.New(t)
-	ctx := common.NewContext()
-	terminator := NewEnvironmentTerminator(ctx, "foo")
-	assert.NotNil(terminator)
-}
-
-type mockedStackManager struct {
+type mockedStackManagerForUpsert struct {
 	mock.Mock
 }
 
-func (m *mockedStackManager) AwaitFinalStatus(stackName string) string {
+func (m *mockedStackManagerForUpsert) AwaitFinalStatus(stackName string) string {
 	args := m.Called(stackName)
 	return args.String(0)
 }
-func (m *mockedStackManager) UpsertStack(stackName string, templateBodyReader io.Reader, stackParameters map[string]string, stackTags map[string]string) error {
+func (m *mockedStackManagerForUpsert) UpsertStack(stackName string, templateBodyReader io.Reader, stackParameters map[string]string, stackTags map[string]string) error {
 	args := m.Called(stackName)
 	return args.Error(0)
 }
-func (m *mockedStackManager) DeleteStack(stackName string) error {
-	args := m.Called(stackName)
-	return args.Error(0)
-}
-func (m *mockedStackManager) FindLatestImageID(pattern string) (string, error) {
+func (m *mockedStackManagerForUpsert) FindLatestImageID(pattern string) (string, error) {
 	args := m.Called()
 	return args.String(0), args.Error(1)
 }
@@ -102,7 +77,7 @@ func TestEnvironmentEcsUpserter(t *testing.T) {
 
 	vpcInputParams := make(map[string]string)
 
-	stackManager := new(mockedStackManager)
+	stackManager := new(mockedStackManagerForUpsert)
 	stackManager.On("AwaitFinalStatus", "mu-cluster-foo").Return(cloudformation.StackStatusCreateComplete)
 	stackManager.On("UpsertStack", "mu-cluster-foo").Return(nil)
 	stackManager.On("FindLatestImageID").Return("ami-00000", nil)
@@ -125,7 +100,7 @@ func TestEnvironmentVpcUpserter(t *testing.T) {
 
 	vpcInputParams := make(map[string]string)
 
-	stackManager := new(mockedStackManager)
+	stackManager := new(mockedStackManagerForUpsert)
 	stackManager.On("AwaitFinalStatus", "mu-vpc-foo").Return(cloudformation.StackStatusCreateComplete)
 	stackManager.On("UpsertStack", "mu-vpc-foo").Return(nil)
 
@@ -159,7 +134,7 @@ environments:
 
 	vpcInputParams := make(map[string]string)
 
-	stackManager := new(mockedStackManager)
+	stackManager := new(mockedStackManagerForUpsert)
 
 	workflow := new(environmentWorkflow)
 	workflow.environment = &config.Environments[0]
@@ -185,44 +160,4 @@ func loadYamlConfig(yamlString string) (*common.Config, error) {
 	}
 
 	return config, nil
-}
-
-func TestNewEnvironmentEcsTerminator(t *testing.T) {
-	assert := assert.New(t)
-
-	workflow := new(environmentWorkflow)
-	workflow.environment = &common.Environment{
-		Name: "foo",
-	}
-
-	stackManager := new(mockedStackManager)
-	stackManager.On("AwaitFinalStatus", "mu-cluster-foo").Return(cloudformation.StackStatusDeleteComplete)
-	stackManager.On("DeleteStack", "mu-cluster-foo").Return(nil)
-
-	err := workflow.environmentEcsTerminator("foo", stackManager, stackManager)()
-	assert.Nil(err)
-
-	stackManager.AssertExpectations(t)
-	stackManager.AssertNumberOfCalls(t, "AwaitFinalStatus", 1)
-	stackManager.AssertNumberOfCalls(t, "DeleteStack", 1)
-}
-
-func TestNewEnvironmentVpcTerminator(t *testing.T) {
-	assert := assert.New(t)
-
-	workflow := new(environmentWorkflow)
-	workflow.environment = &common.Environment{
-		Name: "foo",
-	}
-
-	stackManager := new(mockedStackManager)
-	stackManager.On("AwaitFinalStatus", "mu-vpc-foo").Return(cloudformation.StackStatusDeleteComplete)
-	stackManager.On("DeleteStack", "mu-vpc-foo").Return(nil)
-
-	err := workflow.environmentVpcTerminator("foo", stackManager, stackManager)()
-	assert.Nil(err)
-
-	stackManager.AssertExpectations(t)
-	stackManager.AssertNumberOfCalls(t, "AwaitFinalStatus", 1)
-	stackManager.AssertNumberOfCalls(t, "DeleteStack", 1)
 }
