+++ workflows/service_push.go
@@ -5,10 +5,6 @@ import (
 	"github.com/stelligent/mu/templates"
 )
 
-type serviceWorkflow struct {
-	service *common.Service
-}
-
 // NewServicePusher create a new workflow for pushing a service to a repo
 func NewServicePusher(ctx *common.Context, tag string) Executor {
 
@@ -20,27 +16,6 @@ func NewServicePusher(ctx *common.Context, tag string) Executor {
 		workflow.serviceBuilder(tag),
 		workflow.servicePusher(tag),
 	)
-}
-
-// NewServiceDeployer create a new workflow for deploying a service in an environment
-func NewServiceDeployer(ctx *common.Context, environmentName string, tag string) Executor {
-
-	workflow := new(serviceWorkflow)
-
-	return newWorkflow(
-		workflow.serviceLoader(&ctx.Config),
-		workflow.serviceDeployer(environmentName, tag),
-	)
-}
-
-// Find a service in config, by name and set the reference
-func (workflow *serviceWorkflow) serviceLoader(config *common.Config) Executor {
-	return func() error {
-		workflow.service = &config.Service
-
-		log.Debugf("Working with service '%s' version '%s'", workflow.service.Name, workflow.service.Revision)
-		return nil
-	}
 }
 
 func (workflow *serviceWorkflow) serviceRepoUpserter(stackUpserter common.StackUpserter, stackWaiter common.StackManager) Executor {
@@ -79,14 +54,6 @@ func (workflow *serviceWorkflow) servicePusher(tag string) Executor {
 	return func() error {
 		service := workflow.service
 		log.Debugf("Pushing service '%s' version '%s' tag '%s'", service.Name, service.Revision, tag)
-		return nil
-	}
-}
-
-func (workflow *serviceWorkflow) serviceDeployer(environmentName string, tag string) Executor {
-	return func() error {
-		service := workflow.service
-		log.Debugf("Deploying service '%s' version '%s' tag '%s' to environment '%s'", service.Name, service.Revision, tag, environmentName)
 		return nil
 	}
 }
